updated bacterial workflow

+++ scripts/virus/compile_kallisto.sh
@@ -6,8 +6,14 @@ GENOMES_FOLDER=$4
 # Create output folders
 mkdir -p "${INDEX_FOLDER}/${METHOD}/genomes"
 
+line_num=0
 # Add genomes to the database and remove individual genomes afterwards, saving only the aggregated genome file
 while IFS=$'\t' read -r LINEAGE SEQUENCE col3; do
+    # Skip header line
+    ((line_num++))
+    if [ $line_num -eq 1 ]; then
+        continue
+    fi
     cp "${GENOMES_FOLDER}/${LINEAGE}/${SEQUENCE}" "${INDEX_FOLDER}/${METHOD}/genomes/"
     cat "${INDEX_FOLDER}/${METHOD}/genomes/${SEQUENCE}" >> "${INDEX_FOLDER}/${METHOD}/genomes/all_genomes.fasta"
     rm "${INDEX_FOLDER}/${METHOD}/genomes/${SEQUENCE}"
